Improve error handling & encapsulation of sponsor
