+++ apps/bot/src/buttons/Buttons_Sponsor.res
@@ -21,35 +21,12 @@ let envConfig = switch Env.getConfig() {
 | Error(err) => err->Env.EnvError->raise
 }
 
-let noUnusedSponsorshipsOptions = () =>
-  {
-    "content": "There are no sponsorships available in the Discord pool. Please try again later.",
-    "ephemeral": true,
-  }
-
-let unsuccessfulSponsorMessageOptions = async uuid => {
-  let verifyUrl = `${brightIdLinkVerificationEndpoint}/${uuid}`
-  let uri = `${brightIdAppDeeplink}/${uuid}`
-  let canvas = await makeCanvasFromUri(uri)
-  let attachment = await createMessageAttachmentFromCanvas(canvas)
-  let row = makeBeforeSponsorActionRow("Retry Sponsor", verifyUrl)
-  {
-    "content": "Your sponsor request failed. \n\n This is often due to the BrightID App not being linked to Discord. Please scan this QR code in the BrightID mobile app then retry your sponsorship request.\n\n",
-    "files": [attachment],
-    "ephemeral": true,
-    "components": [row],
-  }
-}
-let sponsorRequestSubmittedMessageOptions = async uuid => {
-  let uri = `${brightIdAppDeeplink}/${uuid}`
-  let canvas = await makeCanvasFromUri(uri)
-  let attachment = await createMessageAttachmentFromCanvas(canvas)
+let sponsorRequestSubmittedMessageOptions = async () => {
   let nowInSeconds = Math.round(Date.now() /. 1000.)
   let fifteenMinutesAfter = 15. *. 60. +. nowInSeconds
-  let content = `You sponsor request has been submitted! \n\n Make sure you have scanned this QR code in the BrightID mobile app to confirm your sponsor and link Discord to BrightID. \n This process will timeout <t:${fifteenMinutesAfter->Float.toString}:R>.\n\nPlease be patient until time expires \n`
+  let content = `You sponsor request has been submitted! \n\n Make sure you have scanned the QR code above in the BrightID mobile app to confirm your sponsor and link Discord to BrightID. \n This process will timeout <t:${fifteenMinutesAfter->Float.toString}:R>.\n\nPlease be patient as the BrightID nodes sync your request \n`
   {
     "content": content,
-    "files": [attachment],
     "ephemeral": true,
   }
 }
@@ -71,19 +48,6 @@ let makeAfterSponsorActionRow = label => {
     ->MessageButton.setStyle("PRIMARY")
 
   MessageActionRow.make()->MessageActionRow.addComponents([verifyButton])
-}
-
-let successfulSponsorMessageOptions = async uuid => {
-  let uri = `${brightIdAppDeeplink}/${uuid}`
-  let canvas = await makeCanvasFromUri(uri)
-  let attachment = await createMessageAttachmentFromCanvas(canvas)
-  let row = makeAfterSponsorActionRow("Assign BrightID Verified Role")
-  {
-    "content": "You have succesfully been sponsored \n\n If you are verified in BrightID you are all done. Click the button below to assign your role.\n\n",
-    "files": [attachment],
-    "ephemeral": true,
-    "components": [row],
-  }
 }
 
 type sponsorship = Sponsorship(BrightId.Sponsorships.t)
@@ -108,120 +72,6 @@ let checkSponsor = async uuid => {
   | (Ok({data}), _) => Sponsorship(data)
   | (_, Ok(error)) => error->Exceptions.BrightIdError->raise
   | (Error(err), _) => err->Json.Decode.DecodeError->raise
-  }
-}
-
-exception HandleSponsorError(string)
-type sponsor = SponsorSuccess(BrightId.Sponsorships.sponsor)
-type handleSponsor =
-  | SponsorshipUsed
-  | RetriedCommandDuring
-  | NoUnusedSponsorships
-  | TimedOut
-
-@raises([HandleSponsorError, Exn.Error, Json.Decode.DecodeError])
-let rec handleSponsor = async (interaction, ~maybeHash=None, ~attempts=30, uuid) => {
-  open Shared.BrightId
-  open Shared.Decode
-  let guildId = interaction->Interaction.getGuild->Guild.getGuildId
-  let secondsBetweenAttempts = 29 //29 seconds between attempts to leave time for timeout message
-  switch attempts {
-  | 0 => TimedOut
-  | _ =>
-    try {
-      let json = await sponsor(
-        ~key=envConfig["sponsorshipKey"],
-        ~context="Discord",
-        ~contextId=uuid,
-      )
-      switch json->Json.decode(Decode_BrightId.Sponsorships.sponsor) {
-      | Ok({hash}) =>
-        let options = await sponsorRequestSubmittedMessageOptions(uuid)
-        let _ = await Interaction.editReply(interaction, ~options, ())
-        Console.log2(
-          `A sponsor request has been submitted`,
-          {"guild": guildId, "contextId": uuid, "hash": hash},
-        )
-        let _ = await CustomMessages.sponsorshipRequested(interaction, uuid, hash)
-        await handleSponsor(interaction, uuid, ~maybeHash=Some(hash), ~attempts=30)
-      | Error(err) => Json.Decode.DecodeError(err)->raise
-      }
-    } catch {
-    | Exn.Error(error) =>
-      try {
-        let brightIdError =
-          JSON.stringifyAny(error)
-          ->Option.map(JSON.parseExn)
-          ->Option.map(Json.decode(_, Decode_BrightId.Error.data))
-
-        switch brightIdError {
-        | None =>
-          HandleSponsorError(
-            "Handle Sponsor Error: There was a problem JSON parsing the error from sponsor()",
-          )->raise
-        | Some(Error(err)) => err->Json.Decode.DecodeError->raise
-        | Some(Ok({errorNum, errorMessage})) =>
-          switch (errorNum, maybeHash) {
-          //No Sponsorships in the Discord App
-          | (38, _) => NoUnusedSponsorships
-          //Sponsorship already assigned
-          | (_, None) => RetriedCommandDuring
-          | (39, Some(_)) =>
-            let Sponsorship({spendRequested}) = await checkSponsor(uuid)
-            if spendRequested {
-              let options = successfulSponsorMessageOptions(uuid)
-              let _ = await Interaction.editReply(interaction, ~options, ())
-              SponsorshipUsed
-            } else {
-              let _ = await sleep(secondsBetweenAttempts * 1000)
-              await handleSponsor(interaction, uuid, ~maybeHash, ~attempts=attempts - 1)
-            }
-          //App authorized before
-          | (45, Some(_)) =>
-            let Sponsorship({spendRequested}) = await checkSponsor(uuid)
-            if spendRequested {
-              let options = successfulSponsorMessageOptions(uuid)
-              let _ = await Interaction.editReply(interaction, ~options, ())
-              SponsorshipUsed
-            } else {
-              let _ = await sleep(secondsBetweenAttempts * 1000)
-              await handleSponsor(interaction, uuid, ~maybeHash, ~attempts=attempts - 1)
-            }
-
-          // Spend Request Submitted
-          | (46, Some(_)) =>
-            let options = await successfulSponsorMessageOptions(uuid)
-            let _ = await interaction->Interaction.editReply(~options, ())
-            SponsorshipUsed
-
-          // Sponsored Request Recently
-          | (47, Some(_)) =>
-            let Sponsorship({spendRequested}) = await checkSponsor(uuid)
-            if spendRequested {
-              let options = successfulSponsorMessageOptions(uuid)
-              let _ = await Interaction.editReply(interaction, ~options, ())
-              SponsorshipUsed
-            } else {
-              let _ = await sleep(secondsBetweenAttempts * 1000)
-              await handleSponsor(interaction, uuid, ~maybeHash, ~attempts=attempts - 1)
-            }
-
-          | _ => HandleSponsorError(errorMessage)->raise
-          }
-        }
-      } catch {
-      | Exceptions.BrightIdError(_) =>
-        let _ = await sleep(secondsBetweenAttempts * 1000)
-        await handleSponsor(interaction, uuid, ~maybeHash, ~attempts=attempts - 1)
-      | Exn.Error(obj) =>
-        switch Exn.message(obj) {
-        | Some(msg) => HandleSponsorError(msg)->raise
-        | None =>
-          Console.error(obj)
-          HandleSponsorError("Handle Sponsor: Unknown Error")->raise
-        }
-      }
-    }
   }
 }
 
@@ -253,6 +103,7 @@ let execute = async interaction => {
           `Buttons_Sponsor: Guild with guildId: ${guildId} not found in gist`,
         )->raise
       | Some(guildData) =>
+        open Services_Sponsor
         let _ = switch await handleSponsor(interaction, uuid) {
         | exception e => e->raise
         | SponsorshipUsed =>
